+++ src/Quests/State/IResetableState.cs
@@ -1,0 +1,7 @@
+﻿namespace QuestFramework.Quests.State
+{
+    public interface IResetableState
+    {
+        void Reset();
+    }
+}